Refactor docker-specific functions from manager to docker
