+++ pages/pages.go
@@ -21,6 +21,7 @@ import (
 	"net/url"
 	"strings"
 
+	"github.com/google/cadvisor/container/docker"
 	httpmux "github.com/google/cadvisor/http/mux"
 	info "github.com/google/cadvisor/info/v1"
 	"github.com/google/cadvisor/manager"
@@ -62,7 +63,7 @@ type pageData struct {
 	Root                   string
 	DockerStatus           []keyVal
 	DockerDriverStatus     []keyVal
-	DockerImages           []manager.DockerImage
+	DockerImages           []docker.DockerImage
 }
 
 func init() {
